+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.6.4
+- Added: Raid Tuning Incoming - 20 May
+- Added: Raid Tuning Incoming - May 19
+
 ## 1.6.3
 - Added: March on Quel'Danas Hall of Fame is Now Full
 
@@ -235,6 +239,7 @@
 - Added external link copy and guild chat sharing for selected posts.
 - Added login toasts for recent posts.
 - Added class section shortcuts for class hotfix posts.
+